+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes/main.tf
@@ -81,11 +81,11 @@ module "eks" {
   )
 
   cluster_addons = {
-    coredns    = {
-      addon_version            = "v1.8.7-eksbuild.3"
+    coredns = {
+      addon_version = "v1.8.7-eksbuild.3"
     }
     kube-proxy = {
-      addon_version            = "v1.24.9-eksbuild.1"
+      addon_version = "v1.24.9-eksbuild.1"
     }
     aws-ebs-csi-driver = {
       service_account_role_arn = aws_iam_role.AmazonEKS_EBS_CSI_DriverRole.arn