+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/redis/versions.tf
@@ -6,3 +6,13 @@
 #
 # usage: create an ElastiCache Redis cache
 #------------------------------------------------------------------------------
+terraform {
+  required_version = "{{ cookiecutter.terraform_required_version }}"
+
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "{{ cookiecutter.terraform_provider_hashicorp_aws_version }}"
+    }
+  }
+}